+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Resolved issues:
 - PI-714 Fix timezone issue for calendar on the Visualizer page
 - PI-731 Bounding box is no longer visible in the Pattern Matching for Citizen Scientist users
 - PI-732 User unable to create PM template
+- PI-730 Fix blank Visualizer page in Pattern Matching - Expert's View part in Citizen Scientist
 - Support for Core API v1.1.x changes
 - PI-739 The Visualizer page not loading before switching tabs